requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.2.0 ; python_version >= '3.6'
cachetools==5.3.0 ; python_version ~= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.0.1
click==8.1.3 ; python_version >= '3.7'
cloudevents==1.9.0
deprecation==2.1.0
flask==2.2.3 ; python_version >= '3.7'
frozenlist==1.3.3 ; python_version >= '3.7'
functions-framework==3.3.0
google-api-core[grpc]==2.11.0 ; python_version >= '3.7'
google-auth==2.16.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-tasks==2.13.0
googleapis-common-protos==1.58.0 ; python_version >= '3.7'
grpc-google-iam-v1==0.12.6 ; python_version >= '3.7'
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0 ; platform_system != 'Windows'
idna==3.4 ; python_version >= '3.5'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
multidict==6.0.4 ; python_version >= '3.7'
openai==0.27.0
packaging==23.0 ; python_version >= '3.7'
proto-plus==1.22.2 ; python_version >= '3.6'
protobuf==4.22.0 ; python_version >= '3.7'
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
rsa==4.9 ; python_version >= '3.6'
setuptools==67.4.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
slack-bolt==1.16.2
slack-sdk==3.20.0 ; python_full_version >= '3.6.0'
tqdm==4.65.0 ; python_version >= '3.7'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
watchdog==2.3.1 ; python_version >= '3.6'
werkzeug==2.2.3 ; python_version >= '3.7'
yarl==1.8.2 ; python_version >= '3.7'
